Added Switch component to Line
To line up with OpenDSS specs more closely, I added an optional breaker
to the Line component. Can be toggled under the "coupling" menu for now,
but that can easily be moved wherever.

Internally, the placements have been worked on a bit so the ports get
pushed left to make space, if a breaker is added, and moved if a breaker
is removed. It should just work